requirements.txt
conda==4.8.1
conda-build==3.10.5
Cython==0.29.14
easydict==1.9
h5py==2.10.0
imageio==2.9.0
imagesize==1.2.0
json5==0.8.5
matplotlib==3.3.4
numpy==1.19.5
numpydoc==0.8.0
opencv-python==4.1.2.30
Pillow==8.2.0
pytorch-ignite==0.1.2
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.5.4
tensorboard==2.1.0
tensorboardX==1.8
tensorflow==2.1.0
tensorflow-estimator==2.1.0
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tqdm==4.41.1